change all double to decimal

+++ src/Service.Liquidity.Monitoring.Domain.Models/AssetPortfolioSettings.cs
@@ -6,8 +6,8 @@ namespace Service.Liquidity.Monitoring.Domain.Models
     public class AssetPortfolioSettings
     {
         [DataMember(Order = 1)] public string Asset { get; set; }
-        [DataMember(Order = 2)] public double NetWarningLevel { get; set; }
-        [DataMember(Order = 3)] public double NetDangerLevel { get; set; }
-        [DataMember(Order = 4)] public double NetCriticalLevel { get; set; }
+        [DataMember(Order = 2)] public decimal NetWarningLevel { get; set; }
+        [DataMember(Order = 3)] public decimal NetDangerLevel { get; set; }
+        [DataMember(Order = 4)] public decimal NetCriticalLevel { get; set; }
     }
 }
